rename: home to area page
